Improve code generation by hooking into SWIPL FLI
SWIPL provides a FFI interface called the FLI that can make code
generation more ergonomic and less fallible. This PR uses that
interface. There are some rough edges in the library I used (cc
terminusdb-labs/swipl-rs#4), but it's still a lot better than generating
text and parsing SWIPL's output as text (which would probably not work
if an exception is raised).

There is still a bit more that needs to be done in terms of ergonomics
and error handling, but this prototype is at a pretty good state at the
moment.

This commit is a squash of the following commits:
- [WIP] Improved codegen
- Revamp prelude/builtins, move save state to target/
- Remove unnecessary primitive, unnecessary dependencies, and fix clippy
  warnings
- Undo some parser/deps changes
- Revert the previous
- Remove merge artifact
- Fix code for new AST

Co-authored-by: Caden Haustein <code@brightlysalty.33mail.com>